+++ packages/data/index.js
@@ -1,1 +1,0 @@
-// This is just for TS declarations to work